fix: last_activity_at is nil when conv is created
The payload does not include last_activity_at when the conversation is created.
Becuase of this the frontend is not able to sort the conversations when appending this.
Another problem is that the last_activity_at is not always present, it is added only when a message is created, and it updates it. So this can be nil when the conversation is created, so we fallback to created_at only at the presentation layer